refactor google doc commentary parser file structure
